+++ test/fixtures/browserstack-app/features/support/env.rb
@@ -1,2 +1,6 @@
 $api_key = '12312312312312312312312312312312'
 ENV['BUGSNAG_API_KEY'] = $api_key
+
+AfterConfiguration do |_config|
+  MazeRunner.config.enforce_bugsnag_integrity = false
+end